+++ changelog/unreleased/pull-4708
@@ -1,10 +1,9 @@
 Enhancement: Back up and restore SecurityDescriptors on Windows
 
-Restic did not back up SecurityDescriptors of files on Windows.
-Restic now backs up and restores SecurityDescriptors (which includes owner, group, 
-discretionary access control list (DACL), system access control list (SACL)) 
-when backing up files and folders on Windows. This requires the user to be
-a member of backup operators or the application must be run as admin.
+Restic now backs up and restores SecurityDescriptors when backing up files and folders 
+on Windows which includes owner, group, discretionary access control list (DACL), 
+system access control list (SACL). This requires the user to be a member of backup 
+operators or the application must be run as admin.
 If that is not the case, only the current user's owner, group and DACL will be backed up 
 and during restore only the DACL of the backed file will be restored while the current 
 user's owner and group will be set during the restore.
